refactor(scanner): extract strategy/tune.py — _classify_tune_result + migrate script + test (#225 PR3)
